+++ instrumentation/racecar/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Release History: opentelemetry-instrumentation-racecar
+
+### v0.1.0 / 2022-11-08
+
+* Initial release.